Move check_for_missing_values up to abc/processor.py
Apply in calculator and timestamp_differ
